Open a new file descriptor for every corpus iteration
The previous implementation of common_types.Corpus used a
single file descriptor for all access. This is not a problem
when populating an in-memory cache (as in this case the file
descriptor is only accessed once). It is a problem for on-disk
storage, though, since it causes inference between multiple
iterations of the corpus. This leads to random jumps between
rows. The new implementation provides an InputFile class that
can safely be iterated over, and an OutputFile class that
features nothing but a write method. Both classes handle
on-the-fly compression and decompression automatically.