Support sparse matrix format in ttm redim
This includes the addition of new methods to Column, that return
some column as either a scipy sparse matrix or as a numpy array.
There is also a general purpose method that autodetects the data
sparsity based on the first few rows. This should allow ttm redim
to scale for lda with huge bow vectors, for instance.