Fix ttm comp combinations order
Document combinations in 'ttm comp' were previously created with
documents sorted in no particular order. Since not all permutations
are created by itertools' combinations, this meant that a combination
of two documents would appear either as A-B or as B-A, depending
on the order of rows in the file under investigation. This lead to
erroneous results for shuffled input files. This commit fixes the
issue by sorting documents before combinations, which ensures that a
combination of documents A and B always appears in the same order. It
will either always appear as A-B or it will always appear as B-A, but
it will not appear in different forms for differently sorted files.