Update version number in setup.py
This bumps the version number to 0.1.0.dev. According to how
setuptools handles version numbers, the '.dev' suffix is a
pre-release tag that is sorted before the corresponding version
number without that tag. I. e. 0.1.0.dev is lower than 0.1.0.
Hence this change basically means that this project has not yet
reached v0.1.0, but that it might reach that version at some point.